requirements.txt
Bootstrap-Flask==1.3.2
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-WTF==0.14.3
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
phonenumbers==8.12.5
PyJWT==1.7.1
python-dotenv==0.13.0
pytz==2020.1
requests==2.23.0
six==1.15.0
twilio==6.41.0
urllib3==1.25.9
Werkzeug==1.0.1
WTForms==2.3.1
